requirements.txt
deepface==0.0.75
Flask==2.1.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
Pillow==9.2.0